added npm scripts
