+++ README.md
@@ -20,10 +20,8 @@ When setting up local environment, do the following:
       
       (in PostgreSQL) Press `ctrl+D` to exit
 
-2. In server/models/index.js, set variable 'resetDB' to true. This will wipe the database, and rebuild the tables.
-3. In server/models/index.js, set variable 'seedDB' to true. This will add the seed data from seedData.json
-4. Run server (nodemone server/server.js). Steps 2 and 3 will happen at server start.
-5. Be sure to set the two variables back to false, otherwise 2 and 3 will happen everytime server is started.
+2. `npm run dbreset` to drop the database and rebuild the tables.
+3. `npm run seed` to seed the database. This takes a while.
 
 
 ###whiteboard photos:
