+++ server/config/middleware.js
@@ -3,6 +3,11 @@ var cookieParser = require('cookie-parser');
 var session = require('express-session');
 var db = require('../models/index');
 var passport = require('../auth/passport');
+//Sync the database initially
+db.sequelize.sync({force: false})
+.then(function(){
+  console.log("Database synced!");
+});
 
 module.exports = function (app, express) {
   //Function for authenticating routes
